% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covdatasim}
\alias{covdatasim}
\title{Correlated Covariates data}
\format{
A dataset with 2000 rows and 5 variables
\describe{
  \item{ID}{Subject ID}
  \item{AGE}{Age in years}
  \item{WT}{Weight in kg}
  \item{Sex}{0=male; 1=female}
  \item{ALB}{Albumin in g/dL}
}
}
\source{
simulated based on a real dataset
}
\usage{
covdatasim
}
\description{
A example dataset used to illustrate multivariate joint covariate effects.
}
\keyword{datasets}
