% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{draw_key}
\alias{draw_key}
\alias{draw_key_hpath}
\title{Horizontal key drawing functions from ggstance in case it is deprecated}
\usage{
draw_key_hpath(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob.
}
\description{
Horizontal key drawing functions from ggstance in case it is deprecated
}
