% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_data.R
\name{get_sample_data}
\alias{get_sample_data}
\title{Get sample dataset}
\usage{
get_sample_data(dataset = "dfall.csv")
}
\arguments{
\item{dataset}{A sample dataset file.}
}
\description{
Get a sample dataset that is included with the package to plot a forest plot.
}
