% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wtage}
\alias{wtage}
\title{Weight Age CDC growth charts data}
\format{
A dataset with 436 rows and 14 variables
\describe{
  \item{Sex}{1=male; 2=female}
  \item{Agemos}{Age in months}
  \item{L}{skewness ditribution parameter}
  \item{M}{location ditribution parameter}
  \item{S}{scale ditribution parameter}
  \item{P3}{}
  \item{P5}{}
  \item{P10}{}
  \item{P25}{}
  \item{P50}{}
  \item{P75}{}
  \item{P90}{}
  \item{P95}{}
  \item{P97}{}
}
}
\source{
CDC website \url{https://www.cdc.gov/growthcharts/data/zscore/wtage.csv}
}
\usage{
wtage
}
\description{
Weight-for-age, 2 to 20 years, LMS parameters and selected smoothed weight percentiles in kilograms, by sex and age.
}
\keyword{datasets}
