% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{expand_modelframe}
\alias{expand_modelframe}
\title{Expand covariate values choices and reference values varying one at a time}
\usage{
expand_modelframe(..., rv, covcol = "covname")
}
\value{
A data.frame with combination of covariates
}
\description{
Expand covariate values choices and reference values varying one at a time
}
