% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prezista}
\alias{prezista}
\title{Prezista Drug Label Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 6 columns.}
\source{
Table 15 from https://aidsinfo.nih.gov/drugs/397/darunavir/28/professional
}
\usage{
prezista
}
\description{
A dataset containing an excerpt from the official Prezista FDA Drug Label
 to help in the app exploration.
}
\keyword{datasets}
