% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_interactiveforestplot.R
\name{run_interactiveforestplot}
\alias{run_interactiveforestplot}
\title{Run the interactiveforestplot application}
\usage{
run_interactiveforestplot()
}
\description{
Run the \code{interactiveforestplot} application.
}
\examples{
if (interactive()) {
  run_interactiveforestplot()
}
}
