% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wtage}
\alias{wtage}
\title{Weight Age CDC growth charts data}
\format{
A dataset with 436 rows and 14 variables
\describe{
  \item{Sex}{1=male; 2=female}
  \item{Agemos}{Age in months}
  \item{L}{skewness ditribution parameter}
  \item{M}{location ditribution parameter}
  \item{S}{scale ditribution parameter}
  \item{P3}{Smoothed third percentile}
  \item{P5}{Smoothed fifth percentile}
  \item{P10}{Smoothed tenth percentile}
  \item{P25}{Smoothed twenty fifth percentile}
  \item{P50}{Smoothed fiftieth percentile}
  \item{P75}{Smoothed seventy fifth percentile}
  \item{P90}{Smoothed ninetieth percentile}
  \item{P95}{Smoothed ninety fifth percentile}
  \item{P97}{Smoothed ninety seventh percentile}
}
}
\source{
CDC website \url{https://www.cdc.gov/growthcharts/data/zscore/wtage.csv}
}
\usage{
wtage
}
\description{
Weight-for-age, 2 to 20 years, LMS parameters and selected smoothed weight percentiles in kilograms, by sex and age.
}
\keyword{datasets}
